test(export-components): ignore
this test is not entirely accurate.
we test exported files against `/src/index.js` which includes a
different set of exports.
also, it tends to be problematic for people who try to run `npm test`
the very first time (hint: it fails). They need to run `npm start` at
least once, but then the root folder is littered with almost empty
files. which is annoying.

hence, until we figure out what and how we want to export, i'm ignoring
this test for now.