<ButtonWithOptions/> - remove internal usages (#2777)
This PR removes internal usages of the deprecated `<ButtonWithOptions/>` component, that has been replaced by the `<DropdownBase/>` component.

### `<EditableSelector/>`

The `<EditableSelector/>` used <ButtonWithOptions.Button/>` internally, with no apparent reason. It was changed to the upgraded `<Button/>`.

### `<StatsWidget/>`

The `<StatsWidget/>` used it pretty heavily, as a way to add a "filter" dropdown. For example:

```jsx
<StatsWidget
	title="Let's see what's going on with your store"
	statistics={statistics}
>
	<StatsWidget.Filter {...ButtonWithOptionsProps}>
		<ButtonWithOptions.Button />
		<ButtonWithOptions.Option key={1} id={1}>
			Look at me!
		</ButtonWithOptions.Option>
	</StatsWidget.Filter>
</StatsWidget>
```

As we don't want to break consumers, a new component API has been introduced, `<StatsWidget.Filter`**`Button`**`/>`, while the `<StatsWidget.Filter/>` has been deprecated. Internally, it uses the `<DropdownBase/>` and its usages is as follows:

```jsx
<StatsWidget
	title="Let's see what's going on with your store"
	statistics={statistics}
>
	<StatsWidget.FilterButton
		dataHook="StatsWidgetFilter"
		initialSelectedId={1}
		options={[
			{ id: 1, value: 'Look at me!' },
		]}
		onSelect={selected => console.log(selected)}
	/>
</StatsWidget>
```

---

#### Issues with the new `<StatsWidget.FilterButton/>`

It uses the new `<TextButton/>` under the hood, which has different styles than the old `<Button/>` with the `'no-border'` theme.

`<TextButton/>` (idle, hover and spacing):

![screen shot 2019-01-16 at 16 30 58](https://user-images.githubusercontent.com/11786506/51255547-2954bf80-19ac-11e9-9859-cc12ed6d23a4.png)
![screen shot 2019-01-16 at 16 31 06](https://user-images.githubusercontent.com/11786506/51255563-2ce84680-19ac-11e9-8f1d-04fada5b600d.png)
![screen shot 2019-01-16 at 16 33 10](https://user-images.githubusercontent.com/11786506/51255728-79338680-19ac-11e9-9706-921e0504429d.png)

Old `<Button/>` (idle, hover and spacing):

![screen shot 2019-01-16 at 16 31 37](https://user-images.githubusercontent.com/11786506/51255620-48ebe800-19ac-11e9-9ee2-e38a66f3e184.png)
![screen shot 2019-01-16 at 16 31 43](https://user-images.githubusercontent.com/11786506/51255624-4c7f6f00-19ac-11e9-8295-bba4a40070f7.png)
![screen shot 2019-01-16 at 16 33 26](https://user-images.githubusercontent.com/11786506/51255739-7d5fa400-19ac-11e9-990c-afa68a83c63e.png)

I think we should add a new theme to the `<Button/>` (not the `<TextButton/>`, as the new theme should have a similar padding to the `<Button/>` and does behave like one) that would resemble the _4.1a Dropdown + No Border_.

@wuwa @LittleTal @milkyfruit wdyt regarding the theme?