Enable Tooltip testkit access appended content
When using `<Tooltip>`'s `appendTo`, `appendToParent` or `appendByPredicate`, the Tooltip's content is not appended to `document.body` but to the specified parent.
When using Enzyme, the tested component that contains the `<Tooltip>` might be mounted using Enzyme's `mount()`, which doesn't append the component to `docuemnt.body` be default.
In such cases, `getContentRoot` will not find the Tooltip's content under `document.body` and it needs to look for it under the testkit's `wrapper`.